requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
vaderSentiment==3.3.2